move package fragment dsl
