+++ app/src/main/AndroidManifest.xml
@@ -58,7 +58,7 @@
             android:name=".navigation.compose.ComposeNavActivity"
             android:exported="false" />
         <activity
-            android:name=".navigation.fragment.FragmentNavActivity"
+            android:name=".navigation.fragment.dsl.FragmentNavActivity"
             android:exported="false" />
 
         <!-- RecyclerView -->
