+++ app/src/main/java/com/cocoslime/presentation/navigation/fragment/dsl/FragmentNavActivity.kt
@@ -1,5 +1,7 @@
-package com.cocoslime.presentation.navigation.fragment
+package com.cocoslime.presentation.navigation.fragment.dsl
 
+import android.annotation.SuppressLint
+import androidx.navigation.NavController
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
 import androidx.core.view.ViewCompat
@@ -13,6 +15,9 @@ import kotlinx.coroutines.launch
 import com.cocoslime.presentation.R
 import com.cocoslime.presentation.databinding.ActivityFragmentNavBinding
 
+/*
+https://developer.android.com/guide/navigation/design/kotlin-dsl
+ */
 class FragmentNavActivity : FragmentActivity() {
 
     private var _binding: ActivityFragmentNavBinding? = null
@@ -59,7 +64,19 @@ class FragmentNavActivity : FragmentActivity() {
             }
         }
 
+        printBackStack(navController)
+    }
 
+    @SuppressLint("RestrictedApi")
+    private fun printBackStack(navController: NavController) {
+        lifecycleScope.launch {
+            navController.currentBackStack.collect {
+                val routes = it
+                    .map { it.destination }
+                    .joinToString(separator = "\n")
+                "======currentBackStack=======\n${routes}\n==============".also(::println)
+            }
+        }
     }
 
     override fun onDestroy() {
