+++ app/src/main/java/com/cocoslime/presentation/navigation/fragment/dsl/VmDestinationViewModel.kt
@@ -1,4 +1,4 @@
-package com.cocoslime.presentation.navigation.fragment
+package com.cocoslime.presentation.navigation.fragment.dsl
 
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
