+++ app/src/main/res/layout/activity_fragment_nav.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".navigation.fragment.FragmentNavActivity">
+    tools:context=".navigation.fragment.dsl.FragmentNavActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/nav_host_fragment"
